THIS PATCH IS NOT REALLY CORRECT but I have no clues.
On one of my systems (Centos 5.5) PHPdoX guesses the encoding wrongly.
If iconv() sees this 'unknown-8bit' PHP crashes on an unhandled exception,
because unknown 8 bit characters cannot be translated properly.

Probably the right thing would be to catch the error iconv() throws
and then do something ultimatively clever.  In contrast my solution
is to print out unhandled encodings into the log such that it's easy
to add a fix to this file manually.  It works at my side, though.